+++ src/core/Word.hpp
@@ -6,7 +6,7 @@ class Word final : public Element
 {
   public:
     Word() = default;
-    explicit Word(QString descr, int value);
+    explicit Word(QString descr, int16_t value);
     explicit Word(QString descr, int addr, uint8_t low, uint8_t high);
     ~Word() override = default;
 
@@ -14,7 +14,7 @@ class Word final : public Element
     Word &operator=(Word &&) = default;
 
     void setValue(int16_t val) override;
-    int value() const override;
+    int16_t value() const override;
 
     uint8_t low() const noexcept;
     uint8_t high() const noexcept;
